separate rumble and dropshot proto
