+++ api/stats/dropshot.proto
@@ -1,0 +1,59 @@
+syntax = "proto2";
+
+import "api/player_id.proto";
+
+package api.stats;
+
+enum TileState {
+    UNDAMAGED = 0;
+    DAMAGED = 1;
+    DESTROYED = 2;
+}
+
+message DropshotTile {
+    optional int32 id = 1;
+    optional TileState state = 2;
+}
+
+message DropshotTileDamageStat {
+    optional int32 id = 1;
+    optional int32 total_damage = 2;
+}
+
+message DropshotTileStats {
+    repeated DropshotTileDamageStat damage_stats = 1;
+    optional int32 damaged_tiles = 2;
+    optional int32 destroyed_tiles = 3;
+}
+
+
+message DropshotStats {
+    optional int32 total_damage = 1;
+    optional float damage_efficiency = 2;
+}
+
+message DropshotBallPhaseStats {
+    optional int32 phase = 1;
+    optional float average = 2;
+    optional float max = 3;
+    optional float total = 4;
+}
+
+message DropshotDamageEvent {
+    optional api.PlayerId player_id = 1;
+    optional int32 frame_number = 2;
+    optional int32 tile_hit = 3;
+    repeated DropshotTile tiles = 4;
+}
+
+message DropshotBallPhaseChangeEvent {
+    optional int32 frame_number = 1;
+    optional int32 ball_phase = 2;
+    optional api.PlayerId player_id = 3;
+}
+
+message DropshotGameStats {
+    repeated DropshotDamageEvent damage_events = 1;
+    repeated DropshotBallPhaseChangeEvent ball_phase_events = 2;
+    optional DropshotTileStats tile_stats = 3;
+}
