+++ carball/analysis/stats/rumble/rumble.py
@@ -8,7 +8,7 @@ from carball.generated.api.player_pb2 import Player
 from carball.generated.api.stats.player_stats_pb2 import PlayerStats
 from carball.generated.api.stats.team_stats_pb2 import TeamStats
 from carball.generated.api.stats.events_pb2 import RumbleItemEvent
-from carball.generated.api.stats.extra_mode_stats_pb2 import PowerUp, RumbleStats
+from carball.generated.api.stats.rumble_pb2 import PowerUp, RumbleStats
 from carball.json_parser.game import Game
 from carball.analysis.stats.stats import BaseStat
 from carball.generated.api.metadata.game_metadata_pb2 import RANKED_RUMBLE, UNRANKED_RUMBLE
