+++ carball/tests/stats/rumble_test.py
@@ -4,7 +4,7 @@ from carball.analysis.analysis_manager import AnalysisManager
 
 from carball.tests.utils import run_analysis_test_on_replay, get_raw_replays
 
-from carball.generated.api.stats.extra_mode_stats_pb2 import *
+from carball.generated.api.stats.rumble_pb2 import *
 
 
 class Test_Rumble():
